requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0

# Audio Processing
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
pyaudio==0.2.14

# Speech-to-Text (Whisper)
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0

# Text-to-Speech (Piper)
piper-tts==1.2.0

# Utilities
colorama==0.4.6
